Remove sections from components
