+++ src/components/blogcard/index.ts
@@ -1,2 +1,1 @@
 export { BlogCard } from './blog-card';
-export type { Blog } from './blog-card';
