+++ src/components/project/index.ts
@@ -1,2 +1,1 @@
 export { ProjectItem } from './project';
-export type { ProjectType } from './project';
