+++ src/sections/about/about.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/scss//variables';
+@import '../../assets/scss//variables';
 
 .about {
   display: flex;
