+++ src/sections/about/about.tsx
@@ -1,10 +1,8 @@
 import React from 'react';
 import { StaticImage } from 'gatsby-plugin-image';
 
-import { tools } from '../../../assets/icons';
-import { AppLink } from '../../AppLink';
-
-import { useScrollReveal } from '../../../hooks';
+import { tools } from '../../assets/icons';
+import { useScrollReveal } from '../../hooks';
 import './about.scss';
 
 const skills = [
@@ -59,21 +57,20 @@ export const About: React.FC = () => {
         <h2 className="about__heading">
           A little <span>About Me</span>
         </h2>
+
         <p>
           I started Software Development in 2015 as a hobby, just out of
           curiosity after reading that video games are made from computer
           programming.
         </p>
+
         <p>
           I&apos;ve been working professionally as a Software Developer for the
-          last ~4 years. I&apos;m currently contracting as a Frontend Engineer
-          with{' '}
-          <AppLink href="https://neighborhoods.com" clearStyles>
-            Neighborhoods.com™
-          </AppLink>{' '}
-          (a real estate resource company in Chicago).
+          last ~5 years.
         </p>
+
         <p>Here are some technologies I enjoy working with:</p>
+
         <ul className="tools">
           {skills.map((skill) => (
             <li key={skill.name} className="tools__tool">
