+++ src/sections/hero/hero.scss
@@ -1,4 +1,4 @@
-@import '../../../assets/scss/variables';
+@import '../../assets/scss/variables';
 
 .hero {
   display: flex;
@@ -46,6 +46,7 @@
       margin-bottom: 2rem;
       width: 100%;
     }
+
     &__lead-text,
     &__tag-text {
       font-size: $font-size-heading;
