+++ src/sections/hero/index.ts
